move _results cache to dataset
